Title: 올바른 등식 만들기, Time: 139ms, Memory: 10MB, Status: Passed - Codetree

+++ 251202/올바른 등식 만들기/right-equality.java
@@ -10,7 +10,7 @@ public class Main {
         }
         // Please write your code here.
 
-        int[][] dp = new int[N][41];
+        long[][] dp = new long[N][41];
 
         if (numbers[0] >= -20 && numbers[0] <= 20) {
             dp[0][numbers[0] + 20] += 1;
